Remove dependency to crazyflie_ros by adding msg/srv directly here.
Crazyflie_ros will be deprecated in favor of the Crazyswarm. In order
to avoid unnecessary deep dependencies, we move all the custom messages
and services directly into the Crazyswarm project.

This change also avoids to have crazyflie_cpp twice. Now, only the one
that comes with crazyflie_tools is used.